Fix routing bug; Fix inactive API indicator style
