+++ frontend/source/index.tsx
@@ -6,7 +6,6 @@ import { FilterBar } from './categories';
 import { SearchBar } from './search';
 import { FilterDialog } from './filtersDialog';
 import { searchPlaces } from "../static/scripts/hbnb.js"
-import { API_root } from "../static/scripts/api_root.js"
 
 
 export interface Filters {
@@ -18,16 +17,6 @@ export interface Filters {
 	price_min?: number
 	bedroom_count?: number
 	bathroom_count?: number
-}
-
-export const api_routes = {
-	home: API_root + "status",
-	cities: API_root + "cities",
-	states: API_root + "states",
-	amenities: API_root + "amenities",
-	categories: API_root + "categories",
-	places: API_root + "places",
-	places_search: API_root + "places_search",
 }
