+++ frontend/static/scripts/api_root.js
@@ -4,7 +4,7 @@
  * port. //TODO: Automate this.
  **/
 // Hosted via Docker:
-// export const API_root = 'http://localhost:8000/api/'
+export const API_root = 'http://localhost:8000/api/'
 
 // Hosted via Flask development server (make sure the port matches):
-export const API_root = 'http://localhost:5002/api/v1/'
+// export const API_root = 'http://localhost:5002/api/v1/'
