+++ frontend/static/styles/header.css
@@ -96,6 +96,10 @@ body>header #logo img {
     width: 1.5rem;
 }
 
+.api_center {
+    display: none;
+}
+
 #api_status.available:hover .api_center {
     transition-property: width, height, color;
     transition-duration: .5s;
